chore(cubesql): Rewrite engine first steps -- polish code structure before merge
